requirements.txt
colorama==0.4.6
pillow==10.4.0
